Correct cast. (Shouldn't matter.)
git-svn-id: https://sc2.svn.sourceforge.net/svnroot/sc2/trunk/sc2@3201 8092fc87-c524-0410-9efc-e669fe64eaf9

+++ src/sc2code/netplay/packetsenders.c
@@ -124,7 +124,7 @@ sendFleet(NetConnection *conn, NetplaySide side, const BYTE *ships,
 	packet = Packet_Fleet_create(side, numShips);
 
 	for (i = 0; i < numShips; i++) {
-		packet->ships[i].index = (uint32) i;
+		packet->ships[i].index = (uint8) i;
 		packet->ships[i].ship = ships[i];
 	}
